Make restart after app crash resume last conversation.
The Whisperer now resets the content ID only when ending a conversation. So if the app crashes while a conversation is in progress, that same content ID will be reused when the Whisperer re-enters the conversation. This means the Listener (and the server) don't see the crash; they just think the Whisperer was quiet for a while.

The only issue is that the Whisperer has lost his census of who is in the conversation.