Fix bugs in session save logic.
Apparently apps can become inactive without being put in the background. If they are quit in that situation, there is no warning given. So now whisper sessions "pack their bags" when they go into the background *or* when they become inactive, and they save their session data at that time. So if the app is quit while it's inactive it can be correctly resumed.